Added explicit cobstructors to vector class. Tests compile with GCC and clang again

+++ test/complex_f_base_math_test.cpp
@@ -3,7 +3,7 @@
 
 TEST(Complex_fBaseMathTest, TestUm)
 {
-    double re = 1, im = 2;
+    float re = 1, im = 2;
     GSL::Complex_f z(re, im);
     GSL::Complex_f z2(z.negate());
     EXPECT_EQ(-z, z2);
@@ -11,30 +11,30 @@ TEST(Complex_fBaseMathTest, TestUm)
 
 TEST(Complex_fBaseMathTest, TestIncrementC)
 {
-    double re = 1, im = 2;
+    float re = 1, im = 2;
     GSL::Complex_f z(re, im);
     GSL::Complex_f dz(re/10, im/5);
-    GSL::Complex_f z2(1.1, 2.4);
+    GSL::Complex_f z2(1.1f, 2.4f);
     z += dz;
     EXPECT_EQ(z, z2);
 }
 
 TEST(Complex_fBaseMathTest, TestDecrementC)
 {
-    double re = 1, im = 2;
+    float re = 1, im = 2;
     GSL::Complex_f z(re, im);
     GSL::Complex_f dz(re/10, im/5);
-    GSL::Complex_f z2(0.9, 1.6);
+    GSL::Complex_f z2(0.9f, 1.6f);
     z -= dz;
     EXPECT_EQ(z, z2);
 }
 
 TEST(Complex_fBaseMathTest, TestScaleC)
 {
-    double re = 1, im = 2;
+    float re = 1, im = 2;
     GSL::Complex_f z(re, im);
     GSL::Complex_f dz(re/10, im/5);
-    GSL::Complex_f z2(-0.7, 0.6);
+    GSL::Complex_f z2(-0.7f, 0.6f);
     z *= dz;
     EXPECT_DOUBLE_EQ(z.re(), z2.re());
     EXPECT_DOUBLE_EQ(z.im(), z2.im());
@@ -42,7 +42,7 @@ TEST(Complex_fBaseMathTest, TestScaleC)
 
 TEST(Complex_fBaseMathTest, TestDivC)
 {
-    double re = 1, im = 2;
+    float re = 1, im = 2;
     GSL::Complex_f z(re, im);
     GSL::Complex_f dz(re/10, im/5);
     GSL::Complex_f z2 = z;
@@ -54,30 +54,30 @@ TEST(Complex_fBaseMathTest, TestDivC)
 
 TEST(Complex_fBaseMathTest, TestIncrementR)
 {
-    double re = 1, im = 2;
+    float re = 1, im = 2;
     GSL::Complex_f z(re, im);
-    double dz = 0.1;
-    GSL::Complex_f z2(1.1, 2);
+    float dz = 0.1f;
+    GSL::Complex_f z2(1.1f, 2);
     z += dz;
     EXPECT_EQ(z, z2);
 }
 
 TEST(Complex_fBaseMathTest, TestDecrementR)
 {
-    double re = 1, im = 2;
+    float re = 1, im = 2;
     GSL::Complex_f z(re, im);
-    double dz = 0.1;
-    GSL::Complex_f z2(0.9, 2);
+    float dz = 0.1f;
+    GSL::Complex_f z2(0.9f, 2);
     z -= dz;
     EXPECT_EQ(z, z2);
 }
 
 TEST(Complex_fBaseMathTest, TestScaleR)
 {
-    double re = 1, im = 2;
+    float re = 1, im = 2;
     GSL::Complex_f z(re, im);
-    double dz = 0.1;
-    GSL::Complex_f z2(0.1, 0.2);
+    float dz = 0.1f;
+    GSL::Complex_f z2(0.1f, 0.2f);
     z *= dz;
     EXPECT_DOUBLE_EQ(z.re(), z2.re());
     EXPECT_DOUBLE_EQ(z.im(), z2.im());
@@ -85,9 +85,9 @@ TEST(Complex_fBaseMathTest, TestScaleR)
 
 TEST(Complex_fBaseMathTest, TestDivR)
 {
-    double re = 1, im = 2;
+    float re = 1, im = 2;
     GSL::Complex_f z(re, im);
-    double dz = 0.1;
+    float dz = 0.1f;
     GSL::Complex_f z2 = z;
     z /= dz;
     z2 *= 1/dz;
@@ -97,35 +97,35 @@ TEST(Complex_fBaseMathTest, TestDivR)
 
 TEST(Complex_fBaseMathTest, TestAddC)
 {
-    double re = 1, im = 2;
+    float re = 1, im = 2;
     GSL::Complex_f z(re, im);
     GSL::Complex_f dz(re/10, im/5);
-    GSL::Complex_f z2(1.1, 2.4);
+    GSL::Complex_f z2(1.1f, 2.4f);
     EXPECT_EQ(z+dz, z2);
 }
 
 TEST(Complex_fBaseMathTest, TestSubC)
 {
-    double re = 1, im = 2;
+    float re = 1, im = 2;
     GSL::Complex_f z(re, im);
     GSL::Complex_f dz(re/10, im/5);
-    GSL::Complex_f z2(0.9, 1.6);
+    GSL::Complex_f z2(0.9f, 1.6f);
     EXPECT_EQ(z-dz, z2);
 }
 
 TEST(Complex_fBaseMathTest, TestMultC)
 {
-    double re = 1, im = 2;
+    float re = 1, im = 2;
     GSL::Complex_f z(re, im);
     GSL::Complex_f dz(re/10, im/5);
-    GSL::Complex_f z2(-0.7, 0.6);
+    GSL::Complex_f z2(-0.7f, 0.6f);
     EXPECT_DOUBLE_EQ((z*dz).re(), z2.re());
     EXPECT_DOUBLE_EQ((z*dz).im(), z2.im());
 }
 
 TEST(Complex_fBaseMathTest, TestDivideC)
 {
-    double re = 1, im = 2;
+    float re = 1, im = 2;
     GSL::Complex_f z(re, im);
     GSL::Complex_f dz(re/10, im/5);
     GSL::Complex_f z2 = z * dz.recipr();
@@ -135,37 +135,37 @@ TEST(Complex_fBaseMathTest, TestDivideC)
 
 TEST(Complex_fBaseMathTest, TestAddR)
 {
-    double re = 1, im = 2;
+    float re = 1, im = 2;
     GSL::Complex_f z(re, im);
-    double dz = 0.1;
-    GSL::Complex_f z2(1.1, 2);
+    float dz = 0.1f;
+    GSL::Complex_f z2(1.1f, 2);
     EXPECT_EQ(z+ dz, z2);
 }
 
 TEST(Complex_fBaseMathTest, TestSubR)
 {
-    double re = 1, im = 2;
+    float re = 1, im = 2;
     GSL::Complex_f z(re, im);
-    double dz = 0.1;
-    GSL::Complex_f z2(0.9, 2);
+    float dz = 0.1f;
+    GSL::Complex_f z2(0.9f, 2);
     EXPECT_EQ(z - dz, z2);
 }
 
 TEST(Complex_fBaseMathTest, TestMulR)
 {
-    double re = 1, im = 2;
+    float re = 1, im = 2;
     GSL::Complex_f z(re, im);
-    double dz = 0.1;
-    GSL::Complex_f z2(0.1, 0.2);
+    float dz = 0.1f;
+    GSL::Complex_f z2(0.1f, 0.2f);
     EXPECT_DOUBLE_EQ((z*dz).re(), z2.re());
     EXPECT_DOUBLE_EQ((z*dz).im(), z2.im());
 }
 
 TEST(Complex_fBaseMathTest, TestDivideR)
 {
-    double re = 1, im = 2;
+    float re = 1, im = 2;
     GSL::Complex_f z(re, im);
-    double dz = 0.1;
+    float dz = 0.1f;
     GSL::Complex_f z2 = z*(1/dz);
     EXPECT_DOUBLE_EQ((z/dz).re(), z2.re());
     EXPECT_DOUBLE_EQ((z/dz).im(), z2.im());
