+++ test/matrix_f_test.cpp
@@ -20,7 +20,7 @@ TEST(Matrix_fTest, TestConstructionInitializerList)
     size_t rows = 2, columns = 2;
     for(size_t i = 0; i < rows; i++){
         for(size_t j = 0; j < columns; j++){
-            EXPECT_FLOAT_EQ(m[i][j], columns*i+j+1);
+            EXPECT_FLOAT_EQ(m[i][j], static_cast<float>(columns*i+j)+1);
         }
     }
 }
@@ -42,7 +42,7 @@ TEST(Matrix_fTest, TestConstructionMove)
     GSL::Matrix_f m1{{1, 2}, {3, 4}}, m2(std::move(m1));
     for(size_t i = 0; i < rows; i++){
         for(size_t j = 0; j < columns; j++){
-            EXPECT_FLOAT_EQ(m2[i][j], columns*i+j+1);
+            EXPECT_FLOAT_EQ(m2[i][j], static_cast<float>(columns*i+j)+1);
         }
     }
 }
